Added the hotkey feature
You can use the Ctrl+Alt+W keystroke to fix and release the window
handle of the recorded window. When you start recording, it is in the
released state, so you can switch between the windows and the SCR will
capture the foreground window. When you click Ctrl+Alt+W it will store
the window handle, so if you switch the foreground window to another
one, it will still record the previous window.