Fix formatting of trouble shooting section
Most of the document rendered nicely as with asciidoctor when we renamed it from .txt to .adoc. However some of the indenting in the trouble shooting section was being lost. This turns each of the Q&A's in the trouble shooting into their own section so it's clearer where one starts and stops. I had to guess a bit.